requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt


sudo systemctl restart postgresql
